refactor(db): separate runtime and store modules
