+++ crates/buzz-db/src/store/allowlist.rs
