+++ crates/buzz-db/src/store/deletion.rs
