+++ crates/buzz-db/src/store/event.rs
