+++ crates/buzz-db/src/store/partition.rs
