+++ crates/buzz-db/src/store/replaceable.rs
