+++ crates/buzz-db/src/store/usage.rs
