+++ crates/buzz-db/src/store/user.rs
